Rename: common 패키지 내 response 패키지 생성 및 구조 분리
