+++ src/main/java/com/newzet/api/common/exception/GlobalExceptionHandler.java
@@ -5,6 +5,9 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
 
+import com.newzet.api.common.response.BasicResponseFormat;
+import com.newzet.api.common.response.ResponseFormat;
+
 import lombok.extern.slf4j.Slf4j;
 
 @RestControllerAdvice
