+++ src/main/java/com/newzet/api/common/response/ResponseCode.java
@@ -1,4 +1,4 @@
-package com.newzet.api.common.exception;
+package com.newzet.api.common.response;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
