+++ src/main/java/com/newzet/api/newsletter/exception/NewsletterNotFoundException.java
@@ -1,7 +1,7 @@
 package com.newzet.api.newsletter.exception;
 
 import com.newzet.api.common.exception.NewzetException;
-import com.newzet.api.common.exception.ResponseCode;
+import com.newzet.api.common.response.ResponseCode;
 
 public class NewsletterNotFoundException extends NewzetException {
